requirements.txt
Django==3.1.7
django-admin-list-filter-dropdown==1.0.3
django-cleanup==5.1.0
Pillow==8.2.0
user-models @ git+https://github.com/reimibeta/django-user-models.git
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
rest-framework-utils @ git+https://github.com/reimibeta/django-rest-framework-utils.git
drf-flex-fields==0.8.9